+++ internal/supervisor/backup.go
@@ -36,6 +36,13 @@ type installationBackupStore interface {
 	GetWebhooks(filter *model.WebhookFilter) ([]*model.Webhook, error)
 }
 
+// BackupProvisioner provisions backup jobs on a cluster.
+type BackupProvisioner interface {
+	TriggerBackup(backupMeta *model.InstallationBackup, cluster *model.Cluster, installation *model.Installation) (*model.S3DataResidence, error)
+	CheckBackupStatus(backupMeta *model.InstallationBackup, cluster *model.Cluster) (int64, error)
+	CleanupBackupJob(backup *model.InstallationBackup, cluster *model.Cluster) error
+}
+
 // BackupSupervisor finds backup pending work and effects the required changes.
 //
 // The degree of parallelism is controlled by a weighted semaphore, intended to be shared with
